requirements.txt
absl-py==0.12.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.12
clang==5.0
click==8.0.4
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
colorama==0.4.5
cycler==0.11.0
dataclasses==0.8
dill==0.3.4
filelock==3.4.1
flatbuffers==1.12
ftfy==6.0.3
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.3
grpcio==1.48.2
h5py==3.1.0
huggingface-hub==0.4.0
idna==3.6
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
littleutils==0.2.2
llvmlite==0.36.0
Markdown==3.3.7
matplotlib==3.3.4
numba==0.53.1
numpy==1.19.5
oauthlib==3.2.2
ogb==1.3.6
opencv-python==4.5.4.60
opt-einsum==3.3.0
outdated==0.2.2
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
prettytable==2.5.0
promise==2.3
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-ranger==0.1.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
sacremoses==0.0.53
scikit-learn==0.24.2
scipy==1.5.4
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-addons==0.14.0
tensorflow-datasets==4.5.2
tensorflow-estimator==2.6.0
tensorflow-metadata==1.2.0
termcolor==1.1.0
threadpoolctl==3.1.0
timm==0.6.12
tokenizers==0.12.1
torch-optimizer==0.3.0
tqdm==4.64.1
transformers==4.18.0
typeguard==2.13.3
typing-extensions==3.7.4.3
urllib3==1.26.18
wcwidth==0.2.13
webcolors==1.11.1
Werkzeug==2.0.3
wilds==2.0.0
wrapt==1.12.1
zipp==3.6.0
imageio==2.15.0
